requests　関心を分離
